requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
pillow>=10.0.0
accelerate>=0.20.0
huggingface_hub>=0.20.0


